requirements.txt
# FastAPI Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP and CORS
python-multipart>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.4.0

# Async support
aiofiles>=23.2.1

# Existing research dependencies (from deep_research_from_scratch)
# These should already be installed if the research workflow is working
langgraph>=0.5.4
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-tavily>=0.2.7
langchain-mcp-adapters>=0.1.9
tavily-python>=0.5.0
pydantic>=2.0.0
rich>=14.0.0
